agent: remove picker usage
Instead call grpc.Dial and (*grpc.ClientConn).Close() directly. Picker
has pretty bad problems when call Reset concurrently - it can block
for a long time and then suddenly close connection which was just
created.

So, in this particular case - less magic is better. There are some other
usages of a picker in ca package, but they seem harmless because there is
no calls to Reset now.

Signed-off-by: Alexander Morozov <lk4d4math@gmail.com>